[Feat] ProductService 수정
  - 예외처리 적용

+++ backend/src/main/java/com/example/Surisuri_Masuri/product/service/ProductService.java
@@ -1,6 +1,9 @@
 package com.example.Surisuri_Masuri.product.service;
 
 import com.example.Surisuri_Masuri.common.BaseResponse;
+import com.example.Surisuri_Masuri.exception.EntityException.ManagerException;
+import com.example.Surisuri_Masuri.exception.EntityException.ProductException;
+import com.example.Surisuri_Masuri.exception.ErrorCode;
 import com.example.Surisuri_Masuri.product.model.Product;
 import com.example.Surisuri_Masuri.product.model.dto.request.ProductCreateReq;
 import com.example.Surisuri_Masuri.product.model.dto.request.ProductUpdateReq;
@@ -46,7 +49,9 @@ public class ProductService {
             return BaseResponse.successResponse("상품 검색 성공", res);
         }
 
-        return BaseResponse.successResponse("검색과 일치하는 상품이 존재하지 않습니다", null);
+        else
+            throw new ProductException(ErrorCode.ProductSearch_002,
+                    String.format("존재하지 않는 상품 정보"));
     }
 
     public BaseResponse list(Integer page, Integer size) {
@@ -79,19 +84,25 @@ public class ProductService {
                 product.setPrice(req.getPrice());
             }
             productRepository.save(product);
+
+            return BaseResponse.successResponse("상품 수정 성공", null);
         }
 
-        return BaseResponse.successResponse("상품 수정 성공", null);
+        else throw new ProductException(ErrorCode.ProductUpdate_002, String.format("존재하지 않는 상품 정보"));
+
     }
 
     public BaseResponse delete(Long idx) {
         Optional<Product> result = productRepository.findById(idx);
 
-        Product product = result.get();
+        if(result.isPresent()) {
+            Product product = result.get();
 
-        productRepository.delete(product);
+            productRepository.delete(product);
 
-        return BaseResponse.successResponse("상품 삭제 성공", null);
+            return BaseResponse.successResponse("상품 삭제 성공", null);
+        }
 
+        else throw new ProductException(ErrorCode.ProductDelete_002, String.format("존재하지 않는 상품 정보"));
     }
 }
